refactor(base): remove interface{} and split large files into modules
- Remove ALL interface{} usage:
  * GetStatistics now returns typed EventManagerStats struct
  * FilterRule.Conditions now uses typed FilterConditions struct

- Split observer.go (667→170 lines) into focused modules:
  * metrics.go: OTEL metrics initialization and recording
  * stats.go: Statistics and health status methods
  * ring_buffer_integration.go: Ring buffer integration methods
  * filter_integration.go: Filter integration methods

- Split filter.go (662 lines) into smaller modules:
  * filter_manager.go: Core filter management and file watching
  * filter_compiler.go: Filter compilation for different types
  * filter_config.go: Configuration types with typed conditions

All files now comply with <500 line limit and ZERO interface{} usage.
